requirements.txt
alabaster==0.7.10
appnope==0.1.0
argh==0.26.2
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
boltons==17.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
defusedxml==0.5.0
docutils==0.14
enum34==1.1.6
gitdb2==2.0.2
GitPython==2.1.5
idna==2.6
imagesize==0.7.1
Jinja2==2.9.6
jira==1.0.10
Logbook==1.1.0
MarkupSafe==1.0
marshmallow==2.14.0
mccabe==0.5.3
multi-key-dict==2.0.3
oauthlib==2.0.4
ordereddict==1.1
pathlib2==2.3.0
pathtools==0.1.2
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
PyGithub==1.35
Pygments==2.2.0
PyJWT==1.5.3
python-jenkins==0.4.15
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.3
snowballstemmer==1.2.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
